Fix -Wc++11-narrowing: chrome/
Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2975060
Commit-Queue: Tommy Nyquist <nyquist@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#896283}

+++ chrome/browser/ash/guest_os/vm_sk_forwarding_native_message_host.cc
@@ -47,7 +47,8 @@ const char* const
         "supported.\"}";
 
 // static
-const int VmSKForwardingNativeMessageHost::kOriginCount = base::size(kOrigins);
+const size_t VmSKForwardingNativeMessageHost::kOriginCount =
+    base::size(kOrigins);
 
 // static
 std::unique_ptr<extensions::NativeMessageHost>
