+++ chrome/browser/extensions/api/messaging/native_message_echo_host.h
@@ -31,7 +31,7 @@ class NativeMessageEchoHost : public NativeMessageHost {
  public:
   static const char* const kHostName;
   static const char* const kOrigins[];
-  static const int kOriginCount;
+  static const size_t kOriginCount;
 
   static std::unique_ptr<NativeMessageHost> Create(
       content::BrowserContext* browser_context);
