+++ chrome/browser/ui/app_list/search/search_result_ranker/recurrence_predictor.cc
@@ -8,6 +8,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include <cmath>
 #include <utility>
 
+#include "base/numerics/safe_conversions.h"
 #include "base/time/time.h"
 #include "chrome/browser/ui/app_list/search/search_result_ranker/frecency_store.pb.h"
 #include "chrome/browser/ui/app_list/search/search_result_ranker/histogram_util.h"
@@ -283,8 +284,10 @@ void FrecencyPredictor::FromProto(const RecurrencePredictorProto& proto) {
 
   std::map<unsigned int, TargetData> targets;
   for (const auto& target_data : predictor.targets()) {
-    targets[target_data.id()] = {target_data.last_score(),
-                                 target_data.last_num_updates()};
+    targets[target_data.id()] = {
+        target_data.last_score(),
+        // Proto field defined as uint32.
+        base::saturated_cast<int32_t>(target_data.last_num_updates())};
   }
   targets_.swap(targets);
 }
